Revert "return requests to a single file"
This reverts commit 022d0416a49ee9b0ccf8eb2957fe63435cfca2e4.
